% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_headers}
\alias{map_headers}
\title{Map raw RAIS headers (any dialect) to legacy contracted names}
\usage{
map_headers(raw)
}
\description{
Map raw RAIS headers (any dialect) to legacy contracted names
}
\keyword{internal}
