% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raisqlr-package.R
\docType{package}
\name{raisqlr-package}
\alias{raisqlr}
\alias{raisqlr-package}
\title{raisqlr: Download dos microdados da RAIS e carga em banco de dados PostgreSQL}
\description{
Cataloga, baixa e converte os microdados da RAIS (Relacao Anual de Informacoes Sociais, Ministerio do Trabalho e Emprego) e os carrega em um banco PostgreSQL no esquema contratado 'rais_vinculo_YYYY'/'rais_estabelecimento_YYYY', consumido pelo AEDi e pelo painel PNDR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rodrigoesborges/raisqlr}
  \item Report bugs at \url{https://github.com/rodrigoesborges/raisqlr/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigo.borges@ipea.gov.br}

}
\keyword{internal}
